+++ Moco/Moco/MocoGoal/MocoControlTrackingGoal.h
@@ -66,7 +66,8 @@ public:
     /// already set for the requested control, then the provided weight
     /// replaces the previous weight. An exception is thrown if a weight
     /// for an unknown control is provided.
-    void setWeight(const std::string& controlName, const double& weight) {
+    void setWeightForControl(
+            const std::string& controlName, const double& weight) {
         if (get_control_weights().contains(controlName)) {
             upd_control_weights().get(controlName).setWeight(weight);
         } else {